refactor: Rename AuthScreen to SignInScreen
